# internal/services/gateway/handlers.go
package gateway

import (
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/quantum-suite/platform/internal/domain"
)

// HealthCheck godoc
// @Summary Health check endpoint
// @Description Check if the gateway service is healthy and ready to serve requests
// @Tags health
// @Accept json
// @Produce json
// @Success 200 {object} HealthResponse "Service is healthy"
// @Failure 503 {object} ErrorResponse "Service is unhealthy"
// @Router /health [get]
func (s *Service) HealthCheck(c *gin.Context) {
	// Implementation would check dependencies
	response := HealthResponse{
		Status:    "healthy",
		Version:   "1.0.0",
		Timestamp: "2025-09-06T19:00:00Z",
		Dependencies: []DependencyHealth{
			{Name: "router", Status: "healthy"},
			{Name: "cache", Status: "healthy"},
		},
	}

	c.JSON(http.StatusOK, response)
}

// ReadinessCheck godoc
// @Summary Readiness check endpoint
// @Description Check if the gateway service is ready to serve requests
// @Tags health
// @Accept json
// @Produce json
// @Success 200 {object} HealthResponse "Service is ready"
// @Failure 503 {object} ErrorResponse "Service is not ready"
// @Router /ready [get]
func (s *Service) ReadinessCheck(c *gin.Context) {
	response := HealthResponse{
		Status:    "ready",
		Version:   "1.0.0",
		Timestamp: "2025-09-06T19:00:00Z",
	}

	c.JSON(http.StatusOK, response)
}

// ListModels godoc
// @Summary List available models
// @Description Get a list of all available LLM models from all providers
// @Tags models
// @Accept json
// @Produce json
// @Security BearerAuth
// @Security TenantID
// @Param provider query string false "Filter by provider" Enums(azure-openai,aws-bedrock)
// @Success 200 {object} ModelsResponse "List of available models"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /v1/models [get]
func (s *Service) ListModels(c *gin.Context) {
	// Extract query parameters
	provider := c.Query("provider")

	// Mock response for now
	models := []Model{
		{
			ID:      "gpt-4",
			Object:  "model",
			Created: 1677610602,
			OwnedBy: "azure-openai",
		},
		{
			ID:      "gpt-35-turbo",
			Object:  "model",
			Created: 1677610602,
			OwnedBy: "azure-openai",
		},
		{
			ID:      "claude-3-sonnet",
			Object:  "model",
			Created: 1677610602,
			OwnedBy: "aws-bedrock",
		},
	}

	// Filter by provider if specified
	if provider != "" {
		var filtered []Model
		for _, model := range models {
			if model.OwnedBy == provider {
				filtered = append(filtered, model)
			}
		}
		models = filtered
	}

	response := ModelsResponse{
		Object: "list",
		Data:   models,
	}

	c.JSON(http.StatusOK, response)
}

// CreateChatCompletion godoc
// @Summary Create chat completion
// @Description Generate a chat completion using the specified model and messages
// @Tags completions
// @Accept json
// @Produce json
// @Security BearerAuth
// @Security TenantID
// @Param request body ChatCompletionRequest true "Chat completion request"
// @Success 200 {object} ChatCompletionResponse "Chat completion response"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 429 {object} ErrorResponse "Rate limit exceeded"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /v1/chat/completions [post]
func (s *Service) CreateChatCompletion(c *gin.Context) {
	var req ChatCompletionRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, ErrorResponse{
			Error: ErrorDetail{
				Message: "Invalid request format",
				Type:    "invalid_request_error",
				Code:    "invalid_json",
			},
		})
		return
	}

	// Mock response for now
	response := ChatCompletionResponse{
		ID:      "chatcmpl-123",
		Object:  "chat.completion",
		Created: 1677652288,
		Model:   req.Model,
		Choices: []Choice{
			{
				Index: 0,
				Message: Message{
					Role:    "assistant",
					Content: "Hello! How can I help you today?",
				},
				FinishReason: "stop",
			},
		},
		Usage: Usage{
			PromptTokens:     9,
			CompletionTokens: 12,
			TotalTokens:      21,
		},
	}

	c.JSON(http.StatusOK, response)
}

// CreateCompletion godoc
// @Summary Create completion (legacy)
// @Description Generate a text completion using the specified model (legacy endpoint)
// @Tags completions
// @Accept json
// @Produce json
// @Security BearerAuth
// @Security TenantID
// @Param request body CompletionRequest true "Completion request"
// @Success 200 {object} CompletionResponse "Completion response"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 429 {object} ErrorResponse "Rate limit exceeded"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /v1/completions [post]
func (s *Service) CreateCompletion(c *gin.Context) {
	var req CompletionRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, ErrorResponse{
			Error: ErrorDetail{
				Message: "Invalid request format",
				Type:    "invalid_request_error",
				Code:    "invalid_json",
			},
		})
		return
	}

	// Mock response for now
	response := CompletionResponse{
		ID:      "cmpl-123",
		Object:  "text_completion",
		Created: 1677652288,
		Model:   req.Model,
		Choices: []CompletionChoice{
			{
				Text:         " This is a test response.",
				Index:        0,
				FinishReason: "stop",
			},
		},
		Usage: Usage{
			PromptTokens:     5,
			CompletionTokens: 7,
			TotalTokens:      12,
		},
	}

	c.JSON(http.StatusOK, response)
}

// CreateEmbedding godoc
// @Summary Create embeddings
// @Description Create vector embeddings for the given input text
// @Tags embeddings
// @Accept json
// @Produce json
// @Security BearerAuth
// @Security TenantID
// @Param request body EmbeddingRequest true "Embedding request"
// @Success 200 {object} EmbeddingResponse "Embedding response"
// @Failure 400 {object} ErrorResponse "Bad request"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 429 {object} ErrorResponse "Rate limit exceeded"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /v1/embeddings [post]
func (s *Service) CreateEmbedding(c *gin.Context) {
	var req EmbeddingRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		c.JSON(http.StatusBadRequest, ErrorResponse{
			Error: ErrorDetail{
				Message: "Invalid request format",
				Type:    "invalid_request_error",
				Code:    "invalid_json",
			},
		})
		return
	}

	// Mock response for now
	embeddings := make([]EmbeddingObject, len(req.Input))
	for i := range req.Input {
		// Generate mock 1536-dimensional embedding (typical for OpenAI models)
		embedding := make([]float64, 1536)
		for j := range embedding {
			embedding[j] = 0.1 // Mock values
		}

		embeddings[i] = EmbeddingObject{
			Object:    "embedding",
			Embedding: embedding,
			Index:     i,
		}
	}

	response := EmbeddingResponse{
		Object: "list",
		Data:   embeddings,
		Model:  req.Model,
		Usage: EmbeddingUsage{
			PromptTokens: len(req.Input) * 10, // Rough estimation
			TotalTokens:  len(req.Input) * 10,
		},
	}

	c.JSON(http.StatusOK, response)
}

// GetMetrics godoc
// @Summary Get service metrics
// @Description Get internal metrics for monitoring and debugging
// @Tags internal
// @Accept json
// @Produce json
// @Security BearerAuth
// @Success 200 {object} MetricsResponse "Service metrics"
// @Failure 401 {object} ErrorResponse "Unauthorized"
// @Failure 500 {object} ErrorResponse "Internal server error"
// @Router /v1/internal/metrics [get]
func (s *Service) GetMetrics(c *gin.Context) {
	metrics := MetricsResponse{
		RequestCount: 100,
		ErrorCount:   5,
		Uptime:       "24h30m",
		Version:      "1.0.0",
	}

	c.JSON(http.StatusOK, metrics)
}